requirements.txt
python-telegram-bot==20.8
yt-dlp==2024.3.10
python-dotenv==1.0.0 
pyrogram==2.0.106
tgcrypto==1.2.5
aiohttp==3.9.1
moviepy==1.0.3
numpy==1.26.2
pymongo==4.6.1
google-auth-oauthlib==1.2.0
google-auth==2.27.0
backoff>=2.2.1
google-api-python-client==2.97.0